+++ examples/greeting.rs
@@ -1,46 +1,34 @@
-use std::{
-    env,
-    io::{Read, Write},
-    net::TcpStream,
-    sync::Arc,
-};
+use std::{env, io::Read};
 
-use io_smtp::rfc5321::greeting::{GetSmtpGreeting, GetSmtpGreetingResult};
-use rustls::{ClientConfig, ClientConnection, StreamOwned};
-use rustls_platform_verifier::ConfigVerifierExt;
+use io_smtp::rfc5321::greeting::*;
+use pimalaya_stream::{std::stream::StreamStd, tls::Tls};
 
 fn main() {
     env_logger::init();
 
     let host = env::var("HOST").expect("HOST env var");
-    let port: u16 = env::var("PORT")
+    let port = env::var("PORT")
         .expect("PORT env var")
         .parse()
         .expect("PORT u16");
 
-    let tcp = TcpStream::connect((host.as_str(), port)).unwrap();
-    let server_name = rustls::pki_types::ServerName::try_from(host.clone()).unwrap();
-    let config = ClientConfig::with_platform_verifier().unwrap();
-    let conn = ClientConnection::new(Arc::new(config), server_name).unwrap();
-    let mut stream = StreamOwned::new(conn, tcp);
+    let tls = Tls::default();
+    let mut stream = StreamStd::connect_tls(&host, port, &tls).unwrap();
 
     let mut coroutine = GetSmtpGreeting::new();
-    let mut buf = [0u8; 4096];
-    let mut chunk: Vec<u8>;
     let mut arg: Option<&[u8]> = None;
+    let mut buf = [0u8; 16 * 1024];
 
     let greeting = loop {
         match coroutine.resume(arg.take()) {
             GetSmtpGreetingResult::Ok { greeting, .. } => break greeting,
             GetSmtpGreetingResult::WantsRead => {
                 let n = stream.read(&mut buf).unwrap();
-                chunk = buf[..n].to_vec();
-                arg = Some(&chunk);
+                arg = Some(&buf[..n]);
             }
             GetSmtpGreetingResult::Err(err) => panic!("{err}"),
         }
     };
 
-    let _ = stream.flush();
     println!("greeting: {greeting:#?}");
 }